Add files via upload
You should add the ttf file (which is inside the zipped folder) to your java font library. (otherwise the handwritten font for the menu would not be shown).
just follow this instruction :

To add a TrueType font to Java, just put the .TTF file in the /usr/java/jre/lib/fonts directory.